Print_all done, sep. files for inner and leaf node, exposed bug
